feat: Expose getter for composite world collision filter
This allows the collision filter that is constructed when building or cloning a world to be accessed by users of the world, without the need to rebuild the filter using one of the existing methods.
This is useful for applications which need to do finer-grained collision checking than the `robowflex` API supports directly.